Refactor: move static func to where it belongs
